backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
msgpack==1.1.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.2.0
